Refactor frontend pages
